tests: clean up dumper package integration test struct
Previously, the dumper package stored per-test state within its integration
test struct fields. This is undesirable because local variables make more
sense for mutable state which is only relevant within each subtest method.
This commit cleans that up so it is consistent with how integration tests are
designed in the other packages.

No user-facing functionality changes are present in this commit.